LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/pagereveal-microtask-sequence-expected.txt


FAIL
  View transition promise reactions in incoming page should resolve before
  the rendering continues (rAF, style/layout etc)
 assert_array_equals: expected property 1 to be "updateCallbackDone" but got "ready" (expected array ["pagereveal", "updateCallbackDone", "ready", "rAF", "rAF-microtask", "finished"] got ["pagereveal", "ready", "updateCallbackDone", "rAF", "rAF-microtask", "finished"])
